fix(bundle): extract generateOdeId into utils/odeId.ts to keep browser bundle clean
ElpxImporter imports generateOdeId at runtime, which is bundled into
public/app/yjs/importers.bundle.js for browser use. Importing it from
OdeXmlGenerator pulled in ../constants, which re-exports getIdeviceConfig
from src/services/idevice-config.ts -- a Node-only module using path
and fs-extra. esbuild then failed to bundle for the browser:

  ERROR: Could not resolve "path"
  ERROR: Could not resolve "constants"
  ERROR: Could not resolve "stream"

Move the pure helper into src/shared/export/utils/odeId.ts (no transitive
deps). OdeXmlGenerator re-exports it so existing call sites keep working;
ElpxImporter imports from the new location.
